requirements.txt
streamlit
requests
beautifulsoup4
langchain
faiss-cpu
tiktoken
sentence-transformers
google-generativeai
python-dotenv
